refactor: reuse shared database component in api_public
